Many extensions and assumptions for time-to-full-display
